move to armeria.common
